+++ src/utils/Buffer2D.cpp
@@ -4,7 +4,7 @@ Buffer2D::Buffer2D(unsigned int capacity, unsigned int element_s)
     : max_capacity(capacity), element_size(element_s)
 {
     arr = (Buffer*)malloc(sizeof(Buffer) * capacity);
-    for(unsigned int i = 0; i < max_capacity; i++)
+    for(size_t i = 0; i < max_capacity; i++)
     {
         arr[i] = Buffer(element_size);
     }
